requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
daphne==3.0.2
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.13
django-cors-headers==3.11.0
django-filter==21.1
django-simple-history==3.0.0
django-simpleui==2022.4.9
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
fabric2==2.7.0
filelock==3.6.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
invoke==1.7.0
itypes==1.2.0
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
mysqlclient==2.1.0
nmap==0.0.1
packaging==21.3
paramiko==2.10.3
paramiko-expect==0.3.2
pathlib2==2.3.7.post1
Pillow==9.1.0
pipenv==2022.4.8
platformdirs==2.5.1
prompt-toolkit==3.0.29
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pytz==2022.1
PyYAML==6.0
redis==4.2.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
virtualenv==20.14.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
wrapt==1.14.0
zope.interface==5.4.0
